ENG-4099 implementation inside rest controller

+++ src/main/java/org/entando/kubernetes/controller/digitalexchange/component/EntandoHubRegistryResourceController.java
@@ -18,13 +18,16 @@ import java.util.List;
 import org.entando.kubernetes.model.entandohub.EntandoHubRegistry;
 import org.entando.kubernetes.model.web.response.DeletedObjectResponse;
 import org.entando.kubernetes.model.web.response.SimpleRestResponse;
+import org.entando.kubernetes.security.AuthorizationChecker;
 import org.entando.kubernetes.service.digitalexchange.entandohub.EntandoHubRegistryService;
 import org.entando.kubernetes.validator.EntandoHubRegistryValidator;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -32,12 +35,16 @@ public class EntandoHubRegistryResourceController implements EntandoHubRegistryR
 
     private final EntandoHubRegistryService service;
     private final EntandoHubRegistryValidator validator;
+    private final AuthorizationChecker authorizationChecker;
 
     @Autowired
     public EntandoHubRegistryResourceController(EntandoHubRegistryService service,
-            EntandoHubRegistryValidator validator) {
+            EntandoHubRegistryValidator validator,
+            AuthorizationChecker authorizationChecker) {
+
         this.service = service;
         this.validator = validator;
+        this.authorizationChecker = authorizationChecker;
     }
 
     @Override
@@ -47,7 +54,10 @@ public class EntandoHubRegistryResourceController implements EntandoHubRegistryR
 
     @Override
     public ResponseEntity<SimpleRestResponse<EntandoHubRegistry>> addRegistry(
+            @RequestHeader(HttpHeaders.AUTHORIZATION) String authorizationHeader,
             @RequestBody EntandoHubRegistry entandoHubRegistry) {
+
+        this.authorizationChecker.checkPermissions(authorizationHeader);
         this.validator.validateEntandoHubRegistryOrThrow(entandoHubRegistry, false);
         final EntandoHubRegistry newRegistry = this.service.createRegistry(entandoHubRegistry);
         return new ResponseEntity<>(new SimpleRestResponse<>(newRegistry), HttpStatus.CREATED);
@@ -55,8 +65,10 @@ public class EntandoHubRegistryResourceController implements EntandoHubRegistryR
 
     @Override
     public ResponseEntity<SimpleRestResponse<EntandoHubRegistry>> updateRegistry(
+            @RequestHeader(HttpHeaders.AUTHORIZATION) String authorizationHeader,
             @RequestBody EntandoHubRegistry entandoHubRegistry) {
 
+        this.authorizationChecker.checkPermissions(authorizationHeader);
         this.validator.validateEntandoHubRegistryOrThrow(entandoHubRegistry,
                 EntandoHubRegistryValidator.VALIDATE_ID_TOO);
         final EntandoHubRegistry updatedRegistry = this.service.updateRegistry(entandoHubRegistry);
@@ -64,7 +76,11 @@ public class EntandoHubRegistryResourceController implements EntandoHubRegistryR
     }
 
     @Override
-    public ResponseEntity<SimpleRestResponse<DeletedObjectResponse>> deleteRegistry(@PathVariable(value = "id") String id) {
+    public ResponseEntity<SimpleRestResponse<DeletedObjectResponse>> deleteRegistry(
+            @RequestHeader(HttpHeaders.AUTHORIZATION) String authorizationHeader,
+            @PathVariable(value = "id") String id) {
+
+        this.authorizationChecker.checkPermissions(authorizationHeader);
         final String deleteRegistryName = this.service.deleteRegistry(id);
         return ResponseEntity.ok(new SimpleRestResponse<>(new DeletedObjectResponse(deleteRegistryName)));
     }
